refactor: added default implementation for observe_with and updated atom & atom:undo & reaction & access_state
