+++ src/reactive_state_access/atom.rs
@@ -287,19 +287,6 @@ where
         (opt_previous_value, new_value)
     }
 
-    fn observe_with<F: FnOnce(&T) -> R, R>(&self, func: F) -> R {
-        if let Ok(context) = illicit::get::<RefCell<ReactiveContext>>() {
-            context.borrow_mut().reactive_state_accessors.push(self.id);
-
-            STORE.with(|store_refcell| {
-                store_refcell
-                    .borrow_mut()
-                    .add_dependency(&self.id, &context.borrow().key);
-            });
-        }
-        read_reactive_state_with_id(self.id, func)
-    }
-
     fn id(&self) -> StorageKey {
         self.id
     }
