+++ src/reactive_state_access/state_access.rs
@@ -1,6 +1,6 @@
 use crate::{
-    clone_reactive_state_with_id, hooks_state_functions::*, read_reactive_state_with_id,
-    store::TopoKey, Observable, ReactiveContext,
+    clone_reactive_state_with_id, hooks_state_functions::*, store::TopoKey, Observable,
+    ReactiveContext,
 };
 use std::marker::PhantomData;
 
@@ -211,22 +211,7 @@ where
     }
 
     // <T: 'static, F: FnOnce(&T) -> R, R>(id: StorageKey, func: F) -> R {
-    fn observe_with<F: FnOnce(&T) -> R, R>(&self, func: F) -> R {
-        let id = crate::store::StorageKey::TopoKey(self.id);
-        if let Ok(context) = illicit::get::<RefCell<ReactiveContext>>() {
-            context
-                .borrow_mut()
-                .reactive_state_accessors
-                .push(id.clone());
 
-            STORE.with(|store_refcell| {
-                store_refcell
-                    .borrow_mut()
-                    .add_dependency(&id, &context.borrow().key);
-            });
-        }
-        read_reactive_state_with_id(id, func)
-    }
     fn id(&self) -> StorageKey {
         StorageKey::TopoKey(self.id)
     }
